Update wich fixes redraw issue among other things
- Quartz should now respect if you turn it on/off in the mod list
- New setting: "Force load" makes Quartz ignore any missing/broken nodes
and try to load the skin anyway (basically marks everything as optional)
- Fixed a lot of loading issues which caused Quartz to load multiple
times, not load at all, and try to apply things in weird orders due to
being loaded more than once. This also seems to have fixed the issue
with the purple bar and transperency not working as intended on some
skins.
- Changed the Ctrl+S shortcut to Ctrl+Shift+S to avoid clashes with mods
that already used it before Quartz